+++ NEWS.md
@@ -27,6 +27,9 @@
 
 * Add `edge|int|ext.linewidth` arguments to `gg.inla.mesh` #188
 
+* Add `bru_safe_sp()` to check if `sp` can be used safely (checks `rgdal` availability
+  and `sp` evolution status, optionally forcing use of `sf`) #178
+
 # inlabru 2.7.0
 
 ## Feature overview